only remove NoAV dicoms when field strength does not equal 7

+++ workflow/scripts/symlink_dicoms_by_field_strength.py
@@ -49,9 +49,10 @@ def symlink_dicoms_by_field_strength(source_dicoms_folder: str, output_folder: s
                     os.symlink(os.path.join(session, item), os.path.join(new_folder, item))
 
             #remove dicoms with NoAV in the name
-            for dicom in list(new_folder.rglob('*NoAV*.dcm', recurse_symlinks=True)):
-                if dicom.parent.exists():
-                    os.remove(dicom.parent)
+            if field_value != "7":
+                for dicom in list(new_folder.rglob('*NoAV*.dcm', recurse_symlinks=True)):
+                    if dicom.parent.exists():
+                        os.remove(dicom.parent)
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(
